Allow choosing number of colors in the "Colorful" FX
Allow either choosing how many colors to extract from palettes or choosing to automatically extract all the non-repeating colors from a palette. This allows it to work with custom versions of the C9 palette that include 5 colors (such as the US incandescent standard of Red, Green, Amber, White, Blue) since the 5 color option is no longer hard coded, and is more accurate than attempting to choose 4 or 5 evenly spaced points on the palette. Also expanded the descriptive text for the "Saturation" slider so that it is apparent how to pick colors instead of using the default palette (I didn't even realize it was possible until looking at the source code).